Make sure it passes Windows building

+++ src/hotspot/cpu/x86/assembler_x86.cpp
@@ -10636,7 +10636,7 @@ void Assembler::btq(Register src, int imm8) {
   assert(isByte(imm8), "not a byte");
   InstructionMark im(this);
   int encode = prefixq_and_encode(src->encoding());
-  emit_int16(0x0f, 0xba);
+  emit_int16(0x0f, static_cast<uint8_t>(0xba));
   emit_int8(0xe0|encode);
   emit_int8(imm8);
 }
